requirements.txt
flask
pytest
pillow